changed INTEGER to BIGINT in relations and changed error throwing to dbError.QueryError

+++ src/core/relations/Many-To-Many.ts
@@ -10,8 +10,8 @@ export const createManyToManyRelation = async function (
     const query: string = `CREATE TABLE IF NOT EXISTS "${intermediateTableName}" (
         "createdAt" TIMESTAMP WITH TIME ZONE NOT NULL,
         "updatedAt" TIMESTAMP WITH TIME ZONE NOT NULL,
-        "${tableName}Id" INTEGER REFERENCES "${tableName}" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
-        "${referenceTableName}Id" INTEGER REFERENCES "${referenceTableName}" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
+        "${tableName}Id" BIGINT REFERENCES "${tableName}" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
+        "${referenceTableName}Id" BIGINT REFERENCES "${referenceTableName}" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
         PRIMARY KEY ("${tableName}Id","${referenceTableName}Id")
   )`;
     try {
